Wrap comments in generated SDK
