+++ powershell/resources/templates/longRunningOperationMethod.ejs
@@ -8,7 +8,7 @@
 <%  if(method.parameters&&!method.language.default.pageable?.nextPageOperation) { method.parameters.filter(p => p.implementation != 'Client' && !project.helper.IsConstantParameter(p) && !(p.extensions && p.extensions['x-ms-parameter-grouping'])).forEach(function (parameter) {-%>
 <%# ToDo: should use camel name instead of pascal name here -%>
         /// <param name='<%-parameter.language.default.name%>'>
-        /// <%=parameter.language.default.description%>
+        /// <%=project.helper.wrapComments('        ', '/// ', parameter.language.default.description)%>
         /// </param>
 <% });}-%>
 <% if(!method.language.default.pageable?.nextPageOperation) {-%>
@@ -16,12 +16,12 @@
 <%if(parameter.extensions && parameter.extensions['x-ms-client-flatten'] ) {-%>
 <%project.helper.GetAllPublicVirtualProperties(parameter.schema.language.default.virtualProperties).filter(p => !project.helper.IsConstantEnumProperty(p) && !p.readOnly).forEach(function(vp) {-%>
         /// <param name='<%-vp.property.language.default.name%>'>
-        /// <%= vp.property.language.default.description%>
+        /// <%=project.helper.wrapComments('        ', '/// ', vp.property.language.default.description)%>
         /// </param>
 <%});-%>
 <%} else {-%>
         /// <param name='<%-parameter.language.default.name%>'>
-        /// <%= parameter.language.default.description%>
+        /// <%=project.helper.wrapComments('        ', '/// ', parameter.language.default.description)%>
         /// </param>
 <%}-%>
 <% });-%>
